requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
google-cloud-speech==2.10.0
google-cloud-language==2.3.0
google-cloud-storage
vertexai
python-dotenv

